Move ColorContainer directory to components directory, add propTypes to App.js
