+++ src/components/ColorContainer/ColorContainer.test.js
